Trying a new tack, here. This is a reboot, instead of a refactor, premised upon the assumption that the core responsibility of dry-system is to provide user-friendly strategies for automatically populating a container with an application. As a top-level environmental pattern, the logic underlying the construction of the container should be direct, and clear. I also propose a principle of user intentionality: The user ought to be aware of and responsible for the final state of the container, and where its keys come from, and we are merely offering assistance in making that happen. Conflicts should be surfaced, with attempts to register the same key in multiple ways raising errors, for the user to resolve, instead of silently failing, as now.
This first commit re-implements auto_register. Its behavior is synchronous, instead of being deferred until #finalize. The synchronous behavior allows for the direct realization of the user's intention to turn a directory and module structure into keys registered on their container. The current indirect implementation _mostly_ does that, but allows for some keys to be occluded by other strategies, and also allows for some complex logic around dependencies and bootables, with a second implementation of the auto registration behavior operating in the form of #load_local_component, during the pre-finalization phase. This can be tough to reason about, at multiple levels.

This has implications for integration with Booter and Importer. Those will be addressed in turn, but I want to note upfront that disentangling these concerns, without significant loss of functionality, is of prime concern. Simplicity and elegance will yield to necessity, if it comes to that, but I am of the belief that the system as a whole can be successfully, and radically, decomposed.